refactor: change path from shared/common/types/Fetch to shared/common/types
Fetch folder has been  deleted
