+++ src/modules/account/Profile/services/fetchPutProfileService.ts
@@ -1,6 +1,6 @@
 import api from "shared/services/api";
 import { IFetchPutProfileService } from "modules/account/Profile/services/types/fetchPutProfileService.types";
-import { IFetchResponseDefault } from "shared/common/types/Fetch";
+import { IFetchResponseDefault } from "shared/common/types";
 
 export const fetchPutProfileService = async ({
 	body,
