+++ src/modules/auth/ForgotPassword/services/fetchPostForgotPasswordService.ts
@@ -1,4 +1,4 @@
-import { IFetchResponseDefault } from "shared/common/types/Fetch";
+import { IFetchResponseDefault } from "shared/common/types";
 import api from "shared/services/api";
 import { IFetchPostForgotPasswordService } from "modules/auth/ForgotPassword/services/types";
