+++ src/modules/auth/SignUp/services/SignUp.service.ts
@@ -1,9 +1,12 @@
-import { IFetchResponseDefault } from "shared/common/types/Fetch";
+import { IFetchResponseDefault } from "shared/common/types";
 import api from "shared/services/api";
 import { TypeFetchPostSignUpService } from "../types";
 
 const fetchPostSignUpService = async (payload: TypeFetchPostSignUpService) => {
-	const response = await api.post<IFetchResponseDefault>("auth/create-account", payload);
+	const response = await api.post<IFetchResponseDefault>(
+		"auth/create-account",
+		payload
+	);
 	return response;
 };
