+++ src/modules/auth/SignUp/services/fetchPostSignUpService.ts
@@ -1,4 +1,4 @@
-import { IFetchResponseDefault } from "shared/common/types/Fetch";
+import { IFetchResponseDefault } from "shared/common/types";
 import api from "shared/services/api";
 import { IFetchPostSignUpService } from "modules/auth/SignUp/services/types";
