+++ src/modules/dashboard/Board/components/Dialogs/DialogDeleteOneBoard/services/fetchDeleteOneBoardService.ts
@@ -1,4 +1,4 @@
-import { IFetchResponseDefault } from "shared/common/types/Fetch";
+import { IFetchResponseDefault } from "shared/common/types";
 import api from "shared/services/api";
 
 export const fetchDeleteOneBoardService = async (
