+++ src/modules/dashboard/Board/components/Dialogs/DialogEditBoard/services/types/fetchPatchBoardService.types.ts
@@ -1,5 +1,5 @@
 import { AxiosPromise } from "axios";
-import { IFetchResponseDefault } from "shared/common/types/Fetch";
+import { IFetchResponseDefault } from "shared/common/types";
 
 export interface IFetchPatchBoardService {
 	(
