+++ src/modules/notes/Note/components/Dialogs/DialogDeleteAllNotes/types/DialogDeleteAllNotes.types.ts
@@ -1,7 +1,7 @@
 import { UseMutateFunction } from "@tanstack/react-query";
 import { AxiosResponse } from "axios";
 import { FunctionReturnsVoid } from "shared/common/types";
-import { IFetchResponseDefault } from "shared/common/types/Fetch";
+import { IFetchResponseDefault } from "shared/common/types";
 
 export interface IDialogDeleteAllNotesView {
 	isOpenDialogDeleteAllNotes: boolean;
