+++ src/modules/notes/Note/components/Dialogs/DialogDeleteOneNote/services/fetchDeleteOneNoteService.ts
@@ -1,4 +1,4 @@
-import { IFetchResponseDefault } from "shared/common/types/Fetch";
+import { IFetchResponseDefault } from "shared/common/types";
 import api from "shared/services/api";
 
 export const fetchDeleteOneNoteService = async (noteID: string) => {
