+++ src/modules/notes/NoteEdit/services/fetchPutNoteService.ts
@@ -1,4 +1,4 @@
-import { IFetchResponseDefault } from "shared/common/types/Fetch";
+import { IFetchResponseDefault } from "shared/common/types";
 import api from "shared/services/api";
 import { IFetchPutNoteService } from "modules/notes/NoteEdit/services/types";
