+++ src/modules/notes/NoteEdit/services/types/fetchPostTodosService.types.ts
@@ -1,5 +1,5 @@
 import { ITodo } from "modules/notes/NoteEdit/types";
-import { IFetchResponseDefault } from "shared/common/types/Fetch";
+import { IFetchResponseDefault } from "shared/common/types";
 
 export interface IFetchPostTodosProps {
 	body: {
