+++ src/modules/notes/NoteEdit/services/types/fetchPutNoteService.types.ts
@@ -1,6 +1,6 @@
 import { AxiosPromise } from "axios";
 import { INoteEditForm } from "modules/notes/NoteEdit/types";
-import { IFetchResponseDefault } from "shared/common/types/Fetch";
+import { IFetchResponseDefault } from "shared/common/types";
 
 export interface IFetchPutNoteProps {
 	params: {
