+++ src/modules/texts/Text/components/Dialogs/DialogDeleteAllTexts/types/DialogDeleteAllTexts.types.ts
@@ -1,6 +1,6 @@
 import { UseMutateFunction } from "@tanstack/react-query";
 import { FunctionReturnsVoid } from "shared/common/types";
-import { IFetchResponseDefault } from "shared/common/types/Fetch";
+import { IFetchResponseDefault } from "shared/common/types";
 
 export interface IDialogDeleteAllTextsView {
 	dialogDeleteAllText: { open: boolean; boardID: string };
