+++ src/modules/texts/Text/components/Dialogs/DialogDeleteOneText/services/fetchDeleteOneTextService.ts
@@ -1,4 +1,4 @@
-import { IFetchResponseDefault } from "shared/common/types/Fetch";
+import { IFetchResponseDefault } from "shared/common/types";
 import api from "shared/services/api";
 
 export const fetchDeleteOneTextService = async (textID: string) => {
