+++ src/modules/texts/Text/components/Dialogs/DialogDeleteOneText/types/DialogDeleteOneText.types.ts
@@ -1,6 +1,6 @@
 import { UseMutateFunction } from "@tanstack/react-query";
 import { FunctionReturnsVoid } from "shared/common/types";
-import { IFetchResponseDefault } from "shared/common/types/Fetch";
+import { IFetchResponseDefault } from "shared/common/types";
 
 export interface IDialogDeleteOneTextView {
 	dialogDeleteOneText: { open: boolean; textID: string };
