+++ src/modules/texts/TextEdit/components/DialogShare/services/fetchPatchTextPermissionsService.ts
@@ -1,4 +1,4 @@
-import { IFetchResponseDefault } from "shared/common/types/Fetch";
+import { IFetchResponseDefault } from "shared/common/types";
 import api from "shared/services/api";
 import { IFetchPatchTextPermissionsService } from "modules/texts/TextEdit/components/DialogShare/services/types";
