+++ src/modules/texts/TextEdit/components/DialogShare/types/DialogShare.types.ts
@@ -2,7 +2,7 @@ import { UseMutateFunction } from "@tanstack/react-query";
 import { AxiosResponse } from "axios";
 import { ChangeEvent } from "react";
 import { FunctionReturnsVoid } from "shared/common/types";
-import { IFetchResponseDefault } from "shared/common/types/Fetch";
+import { IFetchResponseDefault } from "shared/common/types";
 
 export interface IData {
 	public: boolean;
