+++ src/modules/texts/TextEdit/components/Markdown/services/fetchPatchTextService.ts
@@ -1,4 +1,4 @@
-import { IFetchResponseDefault } from "shared/common/types/Fetch";
+import { IFetchResponseDefault } from "shared/common/types";
 import api from "shared/services/api";
 import { IFetchPatchTextService } from "modules/texts/TextEdit/components/Markdown/services/types";
