+++ src/modules/texts/TextEdit/components/Markdown/services/types/fetchPatchTextService.types.ts
@@ -1,5 +1,5 @@
 import { IFetchGetOneTextResponse } from "modules/texts/TextEdit/components/Markdown/services/types";
-import { IFetchResponseDefault } from "shared/common/types/Fetch";
+import { IFetchResponseDefault } from "shared/common/types";
 
 export interface IFetchPatchTextProps {
 	params: {
