+++ src/shared/common/types/app-fetch.types.ts
