+++ src/shared/common/types/index.ts
@@ -1,1 +1,2 @@
 export * from "./app.types";
+export * from "./app-fetch.types";
