+++ src/shared/components/Drawer/components/Aside/components/DialogNewBoard/services/fetchPostBoardService.ts
@@ -1,4 +1,4 @@
-import { IFetchResponseDefault } from "shared/common/types/Fetch";
+import { IFetchResponseDefault } from "shared/common/types";
 import api from "shared/services/api";
 import { IFetchPostBoardService } from "shared/components/Drawer/components/Aside/components/DialogNewBoard/services/types";
